Refectored to support vector types (major bugs).
Currently the scalar code still works, and the vector code works when you are modify a vector (e.g. a<-a+1; 0),
but if you try to write to a different vector (c<-a), or simply (a) which writes to the return value, then
arbb crashes with an internal error.  I need to sandbox the functions being called and try to figure out how arbb
is handling global vectors.