removed Reserve method.
Traces always leave enough space for another op to be recorded.
The Commit() method will return false if there is not enough space
to record another op, indicating that the trace should be flushed.

This changes should make it easier to handle multiple traces
without having to check the space left in each trace before
beginning to record.